style: page structe and info
