--- src/components/WeatherDetails.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const WeatherDetails = () => {
-    return (
-        <>
-            <i class="fas fa-wind"></i>
-            <p>3 m/s</p>
-            <p>Wind</p>
-        </>
-    )
-}
-
-export default WeatherDetails;
