+++ src/components/WeatherDetails/style.css
@@ -1,0 +1,22 @@
+.details-info {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 4rem;
+    border-radius: 10px;
+    margin: 0 10px;
+    padding: 10px;
+    background: rgba(255, 255, 255, 0.3);
+    font-size: 1rem;
+    font-weight: bold;
+    line-height: 1.2rem;
+}
+
+.icon {
+    margin-bottom: 5px;
+    font-size: 1.5rem;
+}
+
+.info {
+    font-weight: lighter;
+}
