--- src/components/WeatherInfo.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const WeatherInfo = () => {
-    return (
-        <>
-            <h4>Fri</h4>
-            <section>
-                <i class="fas fa-wind"></i>
-                <p>17º C</p>
-                <p>14º C</p>
-            </section>
-        </>
-    )
-}
-
-export default WeatherInfo;
