+++ src/components/WeatherInfo/index.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import "./style.css";
+
+const WeatherInfo = () => {
+    return (
+        <section className="week-info">
+            <h4>Fri</h4>
+            <section className="temp-info">
+                <i class="fas fa-wind icon"></i>
+                <p>17º C</p>
+                <p className="temp-min">14º C</p>
+            </section>
+        </section>
+    )
+}
+
+export default WeatherInfo;
