+++ src/components/WeatherInfo/style.css
@@ -1,0 +1,31 @@
+.week-info {
+    display: flex;
+    flex-direction: column;
+    margin: 0 10px;
+}
+
+.week-info h4 {
+    font-weight: lighter;
+}
+
+.temp-info {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 4rem;
+    border-radius: 10px;
+    padding: 10px;
+    background: rgba(255, 255, 255, 0.3);
+    font-size: 1rem;
+    font-weight: bold;
+    line-height: 1.2rem;
+}
+
+.icon {
+    margin-bottom: 5px;
+    font-size: 1.5rem;
+}
+
+.temp-min {
+    font-weight: lighter;
+}
